[app/welcome] + directory refactor + service version
