--- src/app/welcome/welcome.component.html
@@ -1,7 +1,0 @@
-<main class="grid min-h-full place-items-center bg-white px-6 py-24 sm:py-32 lg:px-8">
-    <div class="text-center">
-        <h1 class="mt-4 text-3xl font-bold tracking-tight text-gray-900 sm:text-5xl">Hello world!!</h1>
-        <p class="mt-6 text-base leading-7 text-gray-600">Environment: {{envMessage}}</p>
-        <p class="mt-6 text-base leading-7 text-gray-600">App version: {{appVersion}}</p>
-    </div>
-</main>
